more cleanup and refactoring of the ooxml code:1. removed deprecated methods from xssf and interfaces 2. minimized the accessibility of internal constructors 3. more javadocs
git-svn-id: https://svn.apache.org/repos/asf/poi/branches/ooxml@707839 13f79535-47bb-0310-9956-ffa450edef68

+++ src/ooxml/java/org/apache/poi/xssf/usermodel/XSSFCell.java
@@ -383,7 +383,7 @@ public final class XSSFCell implements Cell {
      */
     public XSSFCellStyle getCellStyle() {
         long idx = cell.isSetS() ? cell.getS() : 0;
-        return stylesSource.getStyleAt(idx);
+        return stylesSource.getStyleAt((int)idx);
     }
 
     /**
